+++ CHANGELOG.md
@@ -9,6 +9,11 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Fixed
+
+- Sync: No warning when an explicitly-added CF group is skipped because the profile is not in its
+  compatibility list (#822)
+
 ## [8.6.0] - 2026-04-26
 
 ### Added